+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Changed signature of `CMAES::print_final_info` to take `&[TerminationReason]` instead of `TerminationReason`.
 - Fixed `TolFun` termination criterion checking whether the current generation's function values are all below `tol_fun` instead of whether the range of them is.
 - Added `FunTarget` termination criterion that checks whether a function value threshold has been reached.
+- Added `tol_x_up` option to configure `TolXUp` termination criterion.
 
 ## 0.1.1 (March 4th, 2022)